Prepare first round of input updates for 3.17.

+++ drivers/media/i2c/smiapp-pll.h
@@ -47,7 +47,7 @@ struct smiapp_pll {
 			uint8_t bus_width;
 		} parallel;
 	};
-	uint8_t flags;
+	unsigned long flags;
 	uint8_t binning_horizontal;
 	uint8_t binning_vertical;
 	uint8_t scale_m;
